+++ app/src/main/java/domain/models/TextResult.kt
@@ -3,7 +3,7 @@ package domain.models
 //вероятно не будет использоваться
 data class TextResult(
     val testName: String,
-    val beginPoint: Int,
-    val endPoint: Int,
+    val beginPoint: Double,
+    val endPoint: Double,
     val text: String
 )
