+++ app/src/main/java/domain/usecases/FinishTestUseCase.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.LiveData
 import domain.repository.TestRepository
 
 class FinishTestUseCase (private val testRepository: TestRepository) {
-    fun finishTest(testName: String, point: Int, userName: String): LiveData<String> {
+    fun finishTest(testName: String, point: Double, userName: String): LiveData<String> {
         return testRepository.finishTest(testName, point, userName)
     }
 }
